% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZW_EIC_plot.R
\name{ZZW_EIC_plot}
\alias{ZZW_EIC_plot}
\title{ZZW_EIC_plot}
\usage{
ZZW_EIC_plot(raw.data, pic.name = NULL, rt.range = NULL,
  color = "dodgerblue")
}
\arguments{
\item{raw.data}{The 2 column data.frame. The column_1 represents time, and column_2 represents counts.}

\item{pic.name}{The export picture name. The default is NULL.}

\item{rt.range}{The retention time range.}

\item{color}{The color of line. Default is "dodgerblue".}
}
\description{
EIC plot
}
\author{
Justin Zhou
}
