% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZW_MSMS_plot.R
\name{ZZW_MSMS_plot}
\alias{ZZW_MSMS_plot}
\title{ZZW_MSMS_plot}
\usage{
ZZW_MSMS_plot(exp.spec, std.spec, is.output = FALSE, file.name = NULL,
  dir.path = NULL, pg.format = "pdf")
}
\arguments{
\item{exp.spec}{experimental spectrum: first column: mz; second column:intensity}

\item{std.spec}{standard spectrum: first column: mz; second column:intensity}

\item{is.output}{whether export picture. Default: False}

\item{file.name}{Picture name. Default: "MSMS"}

\item{dir.path}{output path. Default: current directory.}

\item{pg.format}{Output picture format, including "pdf", "png". Default: "pdf"}
}
\description{
Plot MS/MS match plot
}
\examples{
exp.spec <- data.frame(mz=c(78, 96, 107, 134, 158), int=c(1338,74,74,522,98))
std.spec <- data.frame(mz=c(78, 96, 107, 134, 158), int=c(3845,169,258,610,516))
ZZW_MSMS_plot(exp.spec, std.spec, is.output=FALSE)
}
\author{
Zhiwei Zhou
}
