% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWColors.R
\name{ZZWcolors}
\alias{ZZWcolors}
\title{ZZWcolors}
\usage{
ZZWcolors(name = "contrast1")
}
\arguments{
\item{name}{The name of color set. Contrast color: "contrast1", "contrast2", "contrast3", "contrast4", "google". Sequential color: "seq_blue", "seq_red", "seq_green". Default: "contrast1"}
}
\description{
Some colors for plots
}
\author{
Zhiwei Zhou
{zhouzw@sioc.ac.cn}
}
