% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWCylicBarplot.R
\name{ZZWCylicBarplot}
\alias{ZZWCylicBarplot}
\title{ZZWCylicBarplot}
\usage{
ZZWCylicBarplot(raw_data, min_radius = NULL, max_radius = NULL,
  is_output = FALSE, file_name = "cylic_barplot.pdf")
}
\arguments{
\item{raw_data}{data.frame. 2 columns, 1st column: name; 2nd column: frequency}

\item{min_radius}{min radius of cycle}

\item{max_radius}{max radius of cycle}

\item{is.output}{Default: FALSE}
}
\description{
ZZWCylicBarplot
}
\examples{
raw_data <- seq(10,1)
names(raw_data) <- letters[1:10]
raw_data <- data.frame(name=names(raw_data), frequency=as.numeric(raw_data),
                       stringsAsFactors = F)
}
\author{
Zhiwei Zhou
}
