% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWFanPlot.R
\name{ZZWFanPlot}
\alias{ZZWFanPlot}
\title{ZZWFanPlot}
\usage{
ZZWFanPlot(raw_data)
}
\arguments{
\item{raw_data}{a data.frame with different format.\n 1st column: character of type/group, it should be unique value; 2nd: numeric value}
}
\description{
ZZWFanPlot
}
\examples{
type <- c('A','B','C','D','E','F','G')
nums <- c(10,23,8,33,12,40,60)
raw_data <- data.frame(type = type, nums = nums, stringsAsFactors = F)
ZZWFanPlot(raw_data = raw_data)
}
\author{
Zhiwei Zhou
}
