% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWHeatmap.R
\name{ZZWHeatmap}
\alias{ZZWHeatmap}
\title{ZZWHeatmap}
\usage{
ZZWHeatmap(raw_data, group_info, scale = "row", show_rownames = TRUE,
  show_colnames = TRUE, border_color = NA, fontsize_row = 10,
  cluster_rows = TRUE, cluster_cols = TRUE,
  clustering_method = "ward.D")
}
\description{
ZZWHeatmap
}
\author{
Zhiwei Zhou
}
