% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWLineragePlot.R
\name{ZZWLineragePlot}
\alias{ZZWLineragePlot}
\title{ZZWLineragePlot}
\usage{
ZZWLineragePlot(raw.data, x, ymin, ymax, colour = "dodgerblue",
  label.y = "y", label.x = "x", is.output = FALSE,
  file.name = "LineragePlot.pdf", width = 8, height = 6)
}
\arguments{
\item{raw.data}{it should contain 4 columns, "x", "ymin", "ymax", "label"}

\item{colour}{Default: 'dodgerblue'}

\item{label.y}{Default: 'y'}

\item{label.x}{Default: 'x'}

\item{is.output}{Default: FALSE}
}
\description{
ZZWLineragePlot
}
\examples{
mydata <- data.frame(
  Lebal  = c("linerange1","linerange2","linerange3","linerange4","linerange5"),
  x = c(3.5,7,12,16,20),
  ymin   = c(2.5,6.5,3,4.5,3.8),
  ymax   = c(7.5,9.5,9,13.5,4.2),
  label  = c("A","A","A","C","C")
)

ZZWLineragePlot(mydata)
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
