% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWMultiDeckPiePlot.R
\name{ZZWMultiDeckPiePlot}
\alias{ZZWMultiDeckPiePlot}
\title{ZZWMultiDeckPiePlot}
\usage{
ZZWMultiDeckPiePlot(raw_data, mode = c("single", "double", "triple"))
}
\arguments{
\item{raw_data}{a data.frame with different format.\n "single": 1st column: group1, 2nd column: value;  "double": 1st column: group1, 2nd column: group2, 3rd column value; "triple": 1st column: group1, 2nd column: group2, 3rd column: group4, 3rd column: value}

\item{mode}{"single", "double", "triple". Default: "single"}
}
\description{
ZZWMultiDeckPiePlot
}
\examples{
type <- c('A','B','C','D','E','F','G')
type2 <- c(rep('AA',4), rep('BB', 3))
type3 <- c(rep('AAA', 7))
nums <- c(10,23,8,33,12,40,60)
# single mode
raw_data <- data.frame(type = type, nums = nums, stringsAsFactors = F)
# double mode
raw_data <- data.frame(type = type, type2=type2,  nums = nums, stringsAsFactors = F)
# triple mode
raw_data <- data.frame(type = type, type2=type2, type3=type3, nums = nums, stringsAsFactors = F)
}
\author{
Zhiwei Zhou
}
