% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWpieplot.R
\name{ZZWpieplot}
\alias{ZZWpieplot}
\title{ZZWpieplot}
\usage{
ZZWpieplot(raw.data, color = "contrast1", file.name = "Pie plot",
  is.output = TRUE)
}
\arguments{
\item{raw.data}{data for plot}

\item{file.name}{file name. Default: "Pie plot"}

\item{is.output}{Whether output? Default: TRUE}

\item{colors}{please refer \code{\link{ZZWcolors}}}
}
\description{
plot pie plot
}
\examples{
raw.data <- c(1,4,5)
names(raw.data) <- c("group_1", "group_2", "group_3")
ZZWpieplot(raw.data=raw.data, file.name="Pie plot", is.output=F)
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
