% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWPointplot.R
\name{ZZWPointplot}
\alias{ZZWPointplot}
\title{ZZWPointplot}
\usage{
ZZWPointplot(raw_data, x, y, colour, p.shape = 16, p.size = 3,
  p.alpha = 0.5, label.title = "", label.xlab = "x",
  label.ylab = "y", is.plot = FALSE, file.name = "ZZWPointplot.pdf",
  width = 7.5, height = 6.5)
}
\arguments{
\item{raw_data}{data.frame, it should contains 3 columns, 'x', 'y', 'colour'}

\item{p.shape}{shape of point, please ref ggplot2}

\item{p.size}{size of point, please ref ggplot2}

\item{p.alpha}{alpha of point, please ref ggplot2}

\item{label.title}{title, please ref ggplot2}

\item{label.xlab}{xlab, please ref ggplot2}

\item{label.ylab}{ylab, please ref ggplot2}

\item{val.x}{variable x}

\item{val.y}{variable y}

\item{val.colour}{variable colors}
}
\description{
point plot
}
\examples{
raw_data <- data.frame(x=seq(1:10), y=seq(10:1), colour=letters[1:10], stringAsFactors=F)
ZZWPointplot(raw_data, x, y, colour)
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
