% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWTwoLinePlot.R
\name{ZZWTwoLinePlot}
\alias{ZZWTwoLinePlot}
\title{ZZWTwoLinePlot}
\usage{
ZZWTwoLinePlot(data1, data2, x1, y1, x2, y2, sec_axis_scale = c(0, 1800),
  xlab = "x", ylab1 = "y1", ylab2 = "y2")
}
\arguments{
\item{data1}{data set1; It requires to contain 2 columns, x1 and y1}

\item{data2}{data set2; It requires to contain 2 columns, x2 and y2}

\item{sec_axis_scale}{Rescale sec_axis range, it need to fit with major axis. Default: c(0, 1800)}

\item{xlab}{Character.}

\item{ylab1}{Character.}

\item{ylab2}{Character.}
}
\description{
ZZWTwoLinePlot
}
\author{
Zhiwei Zhou
}
