% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWVennPlot.R
\name{ZZWVennPlot}
\alias{ZZWVennPlot}
\title{ZZWVennPlot}
\usage{
ZZWVennPlot(raw_data, height = 6, width = 6,
  colour_type = "contrast1", filename = "ZZWVennPlot.pdf")
}
\arguments{
\item{raw_data}{a list with group name. The group number needs less than 5.}

\item{height}{Default: 6}

\item{colour_type}{Please ref ZZWColors}

\item{filename}{Default: test}

\item{weight}{Default: 6}
}
\description{
ZZWVennPlot
}
\examples{
test <- list(A=1:10,B=3:8,C=6:9)
ZZWVennPlot(raw_data=test)
}
\author{
Zhiwei Zhou
}
