% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWscale.R
\name{ZZWzscore}
\alias{ZZWzscore}
\title{ZZWzcore}
\usage{
ZZWzscore(raw_data, type = c("by_row", "by_column"))
}
\arguments{
\item{raw_data}{a matrix or data.frame in numeric}

\item{type}{"by_column" or "by_row". Default: "by_row"}
}
\description{
ZZWzcore
}
\author{
Zhiwei Zhou
}
