% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWboxplot.R
\name{ZZWboxplot}
\alias{ZZWboxplot}
\title{boxplot}
\usage{
ZZWboxplot(x, y, lable.group, plot.point = T, figure = "jepg", main.name)
}
\arguments{
\item{x, y}{numeric vector}
}
\description{
plot boxplot and plot p
}
\author{
Justin Zhou
}
