% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZWscale.R
\name{ZZWscale}
\alias{ZZWscale}
\title{Scale transformation for numeric matrix as rows.}
\usage{
ZZWscale(raw.data, method = "auto")
}
\arguments{
\item{raw.data}{A numeric matrix or data.frame.}

\item{method}{Scale method: auto, auto scale; "mintmax", min-max scale.}
}
\value{
scaling transformation of \code{raw.data}
}
\description{
This is the function to complete scale transformation fot matrix or data.frame.
}
\details{
it supports auto-scale or min
}
\examples{
test.matrix <- matrix(1:100, ncol=10)
ZZWscale(test.matrix, method = "auto")
}
